Add tramp-file-name-regexp to vc-ignore-dir-regexp
It is an attempt to speed up remote tramp operations.

vc-ignore-dir-regexp is a variable defined in `vc-hooks.el'.

Documentation:
Regexp matching directory names that are not under VC's control.
The default regexp prevents fruitless and time-consuming attempts
to determine the VC status in directories in which filenames are
interpreted as hostnames.